requirements.txt
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
easygui==0.98.3
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
mido==1.3.2
mutagen==1.47.0
numpy==2.0.1
opencv-python==4.10.0.84
packaging==23.2
pillow==10.4.0
pycryptodomex==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.2
websockets==12.0
yt-dlp==2024.8.6
